Update Camera Shake Mechanic
There are two scripts for the camera shake. The first one is on the camera and the script disables itself on start. The second script is on a trigger. This trigger is a child of the Camera object and will activate the camera shake script on the camera when an object enters the trigger area. Once an object leaves the trigger area, the camera shake script is disabled.  Currently the trigger area is just a small box, however, once the horde it created, we can expand the trigger area so that once the horde enters the screen, it will activate the script.

In the test scene, a ball is falling and activates the camera shake script when it enters the trigger area. Once it leaves the trigger area, the camera shake script is once again disabled.